+++ R/cnetplot.R
@@ -31,8 +31,14 @@ setMethod("cnetplot", signature(x = "compareClusterResult"),
 ##' @param circular whether using circular layout
 ##' @param node_label select which labels to be displayed.
 ##' one of 'category', 'gene', 'all' and 'none', default is "all".
-##' @param node_scale scale of node(for "enrichResult" data) or
-##' pie chart(for "compareClusterResult" data)
+##' @param cex_category number indicating the amount by which plotting category
+##' nodes should be scaled relative to the default.
+##' @param cex_gene number indicating the amount by which plotting gene nodes
+##' should be scaled relative to the default.
+##' @param node_label_size size of node label, this parameter has been
+##' changed to cex_label_category and cex_label_gene
+##' @param cex_label_category scale of category node label size
+##' @param cex_label_gene scale of gene node label size
 ##' @importFrom ggraph geom_edge_arc
 ##' @importFrom ggplot2 scale_colour_gradient2
 ##' @author Guangchuang Yu
@@ -43,11 +49,34 @@ cnetplot.enrichResult <- function(x,
                      colorEdge = FALSE,
                      circular = FALSE,
                      node_label = "all",
-                     node_scale = 1,
+                     cex_category = 1,
+                     cex_gene = 1,
+                     node_label_size = NULL,
+                     cex_label_category = 1,
+                     cex_label_gene = 1,
                      ...) {
 
-    node_label <- match.arg(node_label, c("category", "gene", "all", "none"))
+    if (!is.null(node_label_size))
+        message("node_label_size parameter has been changed to 'cex_label_category' and 'cex_label_gene'")
+    # if (is.null(5 * cex_label_category)) {
+        # if (!is.null(node_label_size)) {
+            # 5 * cex_label_category <- node_label_size
+        # } else {
+            # 5 * cex_label_category <- 5
+        # }
+    # }
 
+    # if (is.null(5 * cex_label_gene)) {
+        # if (!is.null(node_label_size)) {
+            # 5 * cex_label_gene <- node_label_size
+        # } else {
+            # 5 * cex_label_gene <- 5
+        # }
+    # }
+
+    label_category <- 5
+    label_gene <- 5
+    node_label <- match.arg(node_label, c("category", "gene", "all", "none"))
     if (circular) {
         layout <- "linear"
         geom_edge <- geom_edge_arc
@@ -63,10 +92,9 @@ cnetplot.enrichResult <- function(x,
 
     size <- sapply(geneSets, length)
     V(g)$size <- min(size)/2
-
     n <- length(geneSets)
     V(g)$size[1:n] <- size
-
+    node_scales <- c(rep(cex_category, n), rep(cex_gene, (length(V(g)) - n)))
     if (colorEdge) {
         E(g)$category <- rep(names(geneSets), sapply(geneSets, length))
         edge_layer <- geom_edge(aes_(color = ~category), alpha=.8)
@@ -78,35 +106,64 @@ cnetplot.enrichResult <- function(x,
         fc <- foldChange[V(g)$name[(n+1):length(V(g))]]
         V(g)$color <- NA
         V(g)$color[(n+1):length(V(g))] <- fc
-        #palette <- fc_palette(fc)
-        p <- ggraph(g, layout=layout, circular = circular) +
-            edge_layer +
-            geom_node_point(aes_(color=~as.numeric(as.character(color)),
-                                 size=~size)) +
-            scale_colour_gradient2(name = "fold change", low = "green",
-                                   mid = "blue", high = "red")
+        show_legend <- c(TRUE, FALSE)
+        names(show_legend) <- c("color", "size")
+        p <- ggraph(g, layout=layout, circular = circular)
+        p <- p + edge_layer +
+            # geom_node_point(aes_(color=~as.numeric(as.character(color)),
+            geom_node_point(aes_(color=~I("#E5C494"), size=~size),
+                data = p$data[1:n, ]) +
+            scale_size(range=c(3, 8) * cex_category) +
+            ggnewscale::new_scale("size") +
+            ggnewscale::new_scale_color() +
+            geom_node_point(aes_(color=~as.numeric(as.character(color)), size=~size),
+                data = p$data[-(1:n), ], show.legend = show_legend) +
+            scale_size(range=c(3, 3) * cex_gene) +
+            scale_colour_gradient2(name = "fold change", low = "blue",
+                                   mid = "white", high = "red")
     } else {
         V(g)$color <- "#B3B3B3"
         V(g)$color[1:n] <- "#E5C494"
-        p <- ggraph(g, layout=layout, circular=circular) +
-            edge_layer +
-            geom_node_point(aes_(color=~I(color), size=~size))
+        p <- ggraph(g, layout=layout, circular=circular)
+        p <- p + edge_layer +
+            geom_node_point(aes_(color=~I(color), size=~size), data = p$data[1:n, ]) +
+            scale_size(range=c(3, 8) * cex_category) +
+            ggnewscale::new_scale("size") +
+            geom_node_point(aes_(color=~I(color), size=~size),
+                data = p$data[-(1:n), ], show.legend = FALSE) +
+            scale_size(range=c(3, 3) * cex_gene)
     }
 
-    p <- p + scale_size(range=c(3, 10) * node_scale,
-                        breaks=unique(round(seq(min(size), max(size), length.out=4)))) +
-        theme_void()
+    p <- p + theme_void()
 
-
+# gan jue zhe li duo chi yi ju, zhi qian yi jing ba bu xu yao de  she zhi cheng le ""
     if (node_label == "category") {
-        p <- p + geom_node_text(aes_(label=~name), data = p$data[1:n,])
+        if (utils::packageVersion("ggrepel") >= "0.9.0") {
+            p <- p + geom_node_text(aes_(label=~name), data = p$data[1:n,],
+                size = label_category * cex_label_category, bg.color = "white")
+        } else {
+            p <- p + geom_node_text(aes_(label=~name), data = p$data[1:n,],
+                size = label_category * cex_label_category)
+        }
     } else if (node_label == "gene") {
-        p <- p + geom_node_text(aes_(label=~name), data = p$data[-c(1:n),], repel=TRUE)
+        if (utils::packageVersion("ggrepel") >= "0.9.0") {
+            p <- p + geom_node_text(aes_(label=~name), data = p$data[-c(1:n),],
+                repel=TRUE, size = label_gene * cex_label_gene, bg.color = "white")
+        } else {
+            p <- p + geom_node_text(aes_(label=~name), data = p$data[-c(1:n),],
+            repel=TRUE, size = label_gene * cex_label_gene)
+        }
     } else if (node_label == "all") {
         if (utils::packageVersion("ggrepel") >= "0.9.0") {
-            p <- p + geom_node_text(aes_(label=~name), repel=TRUE, bg.color = "white")
+            p <- p + geom_node_text(aes_(label=~name), data = p$data[-c(1:n),],
+                    repel=TRUE, size = label_gene * cex_label_gene, bg.color = "white") + 
+                geom_node_text(aes_(label=~name), repel=TRUE,
+                    size = label_category * cex_label_category, bg.color = "white", data = p$data[1:n,]) 
         } else {
-            p <- p + geom_node_text(aes_(label=~name), repel=TRUE)
+            p <- p + geom_node_text(aes_(label=~name), data = p$data[-c(1:n),],
+                    repel=TRUE, size = label_gene * cex_label_gene) + 
+                geom_node_text(aes_(label=~name), data = p$data[1:n,],
+                    repel=TRUE, size = label_category * cex_label_category)
         }
 
     }
@@ -121,8 +178,8 @@ cnetplot.enrichResult <- function(x,
 ##' @param split separate result by 'category' variable
 ##' @param pie proportion of clusters in the pie chart, one of 'equal' (default) or 'Count'
 ##' @param pie_scale scale of pie chart, this parameter has been changed to "node_scale"
-##' @param node_scale scale of pie plot
 ##' @param legend_n number of circle in legend
+##' @param x_loc,y_loc the location of scatterpie legend
 ##' @importFrom ggraph geom_edge_arc
 ##' @noRd
 cnetplot.compareClusterResult <- function(x,
@@ -134,33 +191,66 @@ cnetplot.compareClusterResult <- function(x,
                      node_label = "all",
                      split=NULL,
                      pie = "equal",
-                     node_scale = NULL,
                      pie_scale = NULL,
+                     cex_category = 1,
+                     cex_gene = 1,
                      legend_n = 5,
+                     node_label_size = NULL,
+                     x_loc = NULL,
+                     y_loc = NULL,
+                     cex_label_category = 1,
+                     cex_label_gene = 1,
                      ...) {
 
-    if (!is.null(pie_scale)) message("pie_scale parameter has been changed to 'node_scale'")
+    if (!is.null(node_label_size))
+        message("node_label_size parameter has been changed to 'cex_label_category' and 'cex_label_gene'")
+    # if (is.null(5 * cex_label_category)) {
+        # if (!is.null(node_label_size)) {
+            # cex_label_category <- node_label_size
+        # } else {
+            # cex_label_category <- 2.5
+        # }
+    # }
 
-    if (is.null(node_scale)) {
+    # if (is.null(5 * cex_label_gene)) {
+        # if (!is.null(node_label_size)) {
+            # cex_label_gene <- node_label_size
+        # } else {
+            # cex_label_gene <- 2.5
+        # }
+    # }
+
+    
+    if (!is.null(pie_scale))
+        message("pie_scale parameter has been changed to 'cex_category' and 'cex_gene'")
+    if (is.null(cex_category)) {
         if (!is.null(pie_scale)) {
-            node_scale <- pie_scale
+            cex_category <- pie_scale
         } else {
-            node_scale <- 1
+            cex_category <- 1
         }
     }
 
-
+    if (is.null(cex_gene)) {
+        if (!is.null(pie_scale)) {
+            cex_gene <- pie_scale
+        } else {
+            cex_gene <- 1
+        }
+    }
+    
+    label_category <- 2.5
+    label_gene <- 2.5
+    range_category_size <- c(3, 8)
+    range_gene_size <- c(3, 3)
     y <- fortify(x, showCategory=showCategory,
-                                      includeAll=TRUE, split=split)
+        includeAll=TRUE, split=split)
     y$Cluster <- sub("\n.*", "", y$Cluster)
 
 
-    #n <- update_n(x, showCategory)
-    #y_union <- merge_compareClusterResult(y)
     y_union <- get_y_union(y = y, showCategory = showCategory)
     y <- y[y$ID %in% y_union$ID, ]
     node_label <- match.arg(node_label, c("category", "gene", "all", "none"))
-    ## when y just have one line
 
 
     if (circular) {
@@ -185,10 +275,30 @@ cnetplot.compareClusterResult <- function(x,
         title <- y$Cluster
         p <- ggraph(g, layout=layout, circular=circular)
         p <- p + edge_layer + theme_void() +
-            geom_node_point(aes_(color=~I(color), size=~size)) +
+            # geom_node_point(aes_(color=~I(color), size=~size)) +
+            # labs(title= title) +
+            # scale_size(range=c(3, 8) * mean(node_scales)) + theme(legend.position="none")
+            geom_node_point(aes_(color=~I(color), size=~size),
+                data = p$data[1:n, ]) +
+            scale_size(range = range_category_size * cex_category) +
+            ggnewscale::new_scale("size") +
+            geom_node_point(aes_(color=~I(color), size=~size),
+                data = p$data[-(1:n), ], show.legend = FALSE) +
+            scale_size(range = range_gene_size * cex_gene) +
             labs(title= title) +
-            scale_size(range=c(3, 8) * node_scale) + theme(legend.position="none")+
-            geom_node_text(aes_(label=~name), data = p$data)
+            theme(legend.position="none")
+        if (utils::packageVersion("ggrepel") >= "0.9.0") {
+            p <- p + geom_node_text(aes_(label=~name), data = p$data[-(1:n),],
+                    size = label_gene * cex_label_gene, bg.color = "white", repel=TRUE) + 
+                geom_node_text(aes_(label=~name), data = p$data[1:n,],
+                    size = label_category * cex_label_category, bg.color = "white", repel=TRUE)
+        } else {
+            p <- p + geom_node_text(aes_(label=~name), data = p$data[-(1:n),],
+                    size = label_gene * cex_label_gene, repel=TRUE) + 
+                geom_node_text(aes_(label=~name), data = p$data[1:n,],
+                    size = label_category * cex_label_category, repel=TRUE)
+        }
+
 
         return(p)
     }
@@ -219,11 +329,14 @@ cnetplot.compareClusterResult <- function(x,
     ID_Cluster_mat2$y <- aa$y[ii]
     #add the radius of the pie chart, the radius of go terms mean the number of genes
     ii <- match(rownames(ID_Cluster_mat2)[1:n], y_union$Description)
-    sizee <- sqrt(y_union[ii,9] / sum(y_union[ii,9])) * node_scale
-    ID_Cluster_mat2$radius <- min(sizee)/2
-    ID_Cluster_mat2$radius[1:n] <- sizee
-    x_loc1 <- min(ID_Cluster_mat2$x)
-    y_loc1 <- min(ID_Cluster_mat2$y)
+    node_scales <- c(rep(cex_category, n), rep(cex_gene, (length(V(g)) - n)))
+    sum_yunion <- sum(y_union[ii,9])
+    sizee <- sqrt(y_union[ii,9] / sum_yunion)
+    ## sizee <- sqrt(y_union[ii,9] / sum(y_union[ii,9]))
+    ID_Cluster_mat2$radius <- min(sizee)/2  * sqrt(cex_gene)
+    ID_Cluster_mat2$radius[1:n] <- sizee * sqrt(cex_category)
+    if(is.null(x_loc)) x_loc <- min(ID_Cluster_mat2$x)
+    if(is.null(y_loc)) y_loc <- min(ID_Cluster_mat2$y)
     #node_label
     if (node_label == "category") {
         p$data$name[(n+1):nrow(p$data)] <- ""
@@ -233,6 +346,7 @@ cnetplot.compareClusterResult <- function(x,
         p$data$name <- ""
     }
     if(ncol(ID_Cluster_mat2) > 4) {
+    ## should not have foldChange
         if (!is.null(foldChange)) {
             log_fc <- abs(foldChange)
             genes <- rownames(ID_Cluster_mat2)[(n+1):nrow(ID_Cluster_mat2)]
@@ -246,26 +360,80 @@ cnetplot.compareClusterResult <- function(x,
             gene_fc[is.na(gene_fc)] <- 1
             gene_fc2 <- c(rep(1,n),gene_fc)
             #Assign value to the size of the genes
-            ID_Cluster_mat2$radius <- min(sizee)/2*gene_fc2
-            ID_Cluster_mat2$radius[1:n] <- sizee
+            # ID_Cluster_mat2$radius <- min(sizee)/2*gene_fc2
+            # ID_Cluster_mat2$radius[1:n] <- sizee
+
+            ID_Cluster_mat2$radius <- min(sizee)/2*gene_fc2 * sqrt(cex_gene)
+            ID_Cluster_mat2$radius[1:n] <- sizee * sqrt(cex_category)
+            # p <- p + geom_scatterpie(aes_(x=~x,y=~y,r=~radius),
+                    # data=ID_Cluster_mat2,
+                    # cols=colnames(ID_Cluster_mat2)[1:(ncol(ID_Cluster_mat2)-3)],
+                    # color=NA) +
             p <- p + geom_scatterpie(aes_(x=~x,y=~y,r=~radius),
-                    data=ID_Cluster_mat2,
+                    data=ID_Cluster_mat2[1:n, ],
+                    cols=colnames(ID_Cluster_mat2)[1:(ncol(ID_Cluster_mat2)-3)], color=NA) +
+                geom_scatterpie_legend(ID_Cluster_mat2$radius[1:n],
+                    x=x_loc, y=y_loc + 3, n = legend_n, labeller=function(x) round(x^2 * sum_yunion / cex_category))  +
+                geom_scatterpie(aes_(x=~x,y=~y,r=~radius),
+                    data=ID_Cluster_mat2[-(1:n), ],
                     cols=colnames(ID_Cluster_mat2)[1:(ncol(ID_Cluster_mat2)-3)],
-                    color=NA) +
+                    color=NA, show.legend = FALSE) +
                 coord_equal()+
                 geom_scatterpie_legend(ID_Cluster_mat2$radius[(n+1):nrow(ID_Cluster_mat2)],
-                    x=x_loc1, y=y_loc1, n = legend_n,
-                    labeller=function(x) round(x*2/(min(sizee)),3)) +
-                geom_node_text(aes_(label=~name), repel=TRUE, size=2.5) +
-                theme_void() + labs(fill = "Cluster")
+                    x=x_loc, y=y_loc, n = legend_n,
+                    labeller=function(x) round(x*2/(min(sizee))/sqrt(cex_gene),3)) +
+                ggplot2::annotate("text", x = x_loc + 3, y = y_loc, label = "log2FC")  +
+                ggplot2::annotate("text", x = x_loc + 3, y = y_loc + 3, label = "gene number")
+
+            if (utils::packageVersion("ggrepel") >= "0.9.0") {
+                p <- p + geom_node_text(aes_(label=~name), data = p$data[-(1:n),],
+                        size = label_gene * cex_label_gene, bg.color = "white", repel=TRUE) + 
+                    geom_node_text(aes_(label=~name), data = p$data[1:n,],
+                        size = label_category * cex_label_category, bg.color = "white", repel=TRUE)
+            } else {
+                p <- p + geom_node_text(aes_(label=~name), data = p$data[-(1:n),],
+                        size = label_gene * cex_label_gene, repel=TRUE) + 
+                    geom_node_text(aes_(label=~name), data = p$data[1:n,],
+                        size = label_category * cex_label_category, repel=TRUE)
+            }
+
+
+            p <- p + theme_void() + labs(fill = "Cluster")
             return(p)
         }
-        p <- p + geom_scatterpie(aes_(x=~x,y=~y,r=~radius), data=ID_Cluster_mat2,
+        ## should not have foldChange
+        # p <- p + geom_scatterpie(aes_(x=~x,y=~y,r=~radius), data=ID_Cluster_mat2,
+                # cols=colnames(ID_Cluster_mat2)[1:(ncol(ID_Cluster_mat2)-3)],
+                # color=NA) +
+           # coord_equal()
+        p <- p + geom_scatterpie(aes_(x=~x,y=~y,r=~radius),
+                data=ID_Cluster_mat2[1:n, ],
+                cols=colnames(ID_Cluster_mat2)[1:(ncol(ID_Cluster_mat2)-3)], color=NA) +
+            geom_scatterpie(aes_(x=~x,y=~y,r=~radius),
+                data=ID_Cluster_mat2[-(1:n), ],
                 cols=colnames(ID_Cluster_mat2)[1:(ncol(ID_Cluster_mat2)-3)],
-                color=NA) +
-            coord_equal()+
-            geom_node_text(aes_(label=~name), repel=TRUE, size=2.5) +
-            theme_void() + labs(fill = "Cluster")
+                color=NA, show.legend = FALSE) +
+            coord_equal() +
+            geom_scatterpie_legend(ID_Cluster_mat2$radius[1:n],
+                    x=x_loc, y=y_loc, n = legend_n, labeller=function(x) round(x^2 * sum_yunion / cex_category)) +
+            ggplot2::annotate("text", x = x_loc + 3, y = y_loc, label = "gene number")
+
+        if (utils::packageVersion("ggrepel") >= "0.9.0") {
+            p <- p + geom_node_text(aes_(label=~name), data = p$data[-(1:n),],
+                    size = label_gene * cex_label_gene, bg.color = "white", repel=TRUE) + 
+                geom_node_text(aes_(label=~name), data = p$data[1:n,],
+                    size = label_category * cex_label_category, bg.color = "white", repel=TRUE)
+        } else {
+            p <- p + geom_node_text(aes_(label=~name), data = p$data[-(1:n),],
+                    size = label_gene * cex_label_gene, repel=TRUE) + 
+                geom_node_text(aes_(label=~name), data = p$data[1:n,],
+                    size = label_category * cex_label_category, repel=TRUE)
+        }
+
+
+
+
+        p <- p + theme_void() + labs(fill = "Cluster")
         return(p)
     }
     title <- colnames(ID_Cluster_mat2)[1]
@@ -274,11 +442,30 @@ cnetplot.compareClusterResult <- function(x,
     V(g)$color[1:n] <- "#E5C494"
 
     p <- ggraph(g, layout=layout, circular=circular)
-    p + edge_layer + geom_node_point(aes_(color=~I(color), size=~size)) +
-        labs(title= title) +
-        geom_node_text(aes_(label=~name), data = p$data) +
-        scale_size(range=c(3, 8) * node_scale) + theme_void() +
-        theme(legend.position="none")
+    p <- p + edge_layer +
+        geom_node_point(aes_(color=~I(color), size=~size),
+            data = p$data[1:n, ]) +
+        scale_size(range = range_category_size * cex_category) +
+        ggnewscale::new_scale("size") +
+        geom_node_point(aes_(color=~I(color), size=~size),
+            data = p$data[-(1:n), ], show.legend = FALSE) +
+        scale_size(range = range_gene_size * cex_gene) +
+        labs(title= title)
+
+    if (utils::packageVersion("ggrepel") >= "0.9.0") {
+        p <- p + geom_node_text(aes_(label=~name), data = p$data[-(1:n),],
+                size = label_gene * cex_label_gene, bg.color = "white", repel=TRUE) + 
+            geom_node_text(aes_(label=~name), data = p$data[1:n,],
+                size = label_category * cex_label_category, bg.color = "white", repel=TRUE)
+    } else {
+        p <- p + geom_node_text(aes_(label=~name), data = p$data[-(1:n),],
+                size = label_gene * cex_label_gene, repel=TRUE) + 
+            geom_node_text(aes_(label=~name), data = p$data[1:n,],
+                size = label_category * cex_label_category, repel=TRUE)
+    }
+
+
+    p + theme_void() + theme(legend.position="none")
 }